requirements.txt
opencv-python==4.9.0.80
PyAutoGUI~=0.9.54
numpy~=1.26.4
mss~=9.0.1
pillow~=10.3.0
python-dotenv~=1.0.1
PyGetWindow~=0.0.9